requirements.txt
aiofiles==24.1.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.3
Deprecated==1.2.18
distro==1.9.0
Django==5.1.7
exceptiongroup==1.2.2
get-annotations==0.1.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
netaddr==1.3.0
packaging==25.0
pillow==11.1.0
protobuf==5.29.4
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wrapt==1.17.2
yandex-cloud-ml-sdk==0.9.1
yandexcloud==0.342.0
yookassa==3.5.0
